# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rozgar Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type SignupBodyUserType =
  (typeof SignupBodyUserType)[keyof typeof SignupBodyUserType];

export const SignupBodyUserType = {
  customer: "customer",
  worker: "worker",
} as const;

export interface SignupBody {
  name: string;
  phone: string;
  password: string;
  city: string;
  userType: SignupBodyUserType;
  /** Required for workers */
  skill?: string;
  /** Required for workers (PKR) */
  hourlyRate?: number;
  bio?: string;
}

export type LoginBodyUserType =
  (typeof LoginBodyUserType)[keyof typeof LoginBodyUserType];

export const LoginBodyUserType = {
  customer: "customer",
  worker: "worker",
} as const;

export interface LoginBody {
  phone: string;
  password: string;
  userType: LoginBodyUserType;
}

export type UserInfoUserType =
  (typeof UserInfoUserType)[keyof typeof UserInfoUserType];

export const UserInfoUserType = {
  customer: "customer",
  worker: "worker",
} as const;

export interface UserInfo {
  id: number;
  name: string;
  phone: string;
  city: string;
  userType: UserInfoUserType;
  /** Present for workers */
  workerId?: number;
}

export interface AuthResponse {
  user: UserInfo;
}

export interface WorkerCard {
  id: number;
  userId: number;
  name: string;
  city: string;
  skill: string;
  hourlyRate: number;
  avgRating: number;
  totalReviews: number;
  bio?: string;
}

export interface Review {
  id: number;
  bookingId: number;
  customerId: number;
  customerName?: string;
  workerId: number;
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface WorkerProfile {
  id: number;
  userId: number;
  name: string;
  city: string;
  skill: string;
  hourlyRate: number;
  avgRating: number;
  totalReviews: number;
  bio?: string;
  reviews: Review[];
}

export interface CreateBookingBody {
  workerId: number;
  /** ISO date string */
  date: string;
  /** e.g. 09:00 */
  time: string;
  address: string;
  notes?: string;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface Booking {
  id: number;
  customerId: number;
  customerName?: string;
  workerId: number;
  workerName?: string;
  workerSkill?: string;
  workerHourlyRate?: number;
  date: string;
  time: string;
  address: string;
  notes?: string;
  status: BookingStatus;
  hasReview: boolean;
  createdAt: string;
}

export interface CreateReviewBody {
  bookingId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
}

export type PlatformStatsTopCategoriesItem = {
  category: string;
  count: number;
};

export interface PlatformStats {
  totalWorkers: number;
  totalBookings: number;
  totalCategories: number;
  topCategories: PlatformStatsTopCategoriesItem[];
  cities: string[];
}

export type ListWorkersParams = {
  category?: string;
  city?: string;
  search?: string;
};