# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rozgar Connect API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: workers
    description: Worker listings and profiles
  - name: bookings
    description: Booking management
  - name: reviews
    description: Reviews and ratings
  - name: stats
    description: Platform statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user (customer or worker)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Phone already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: Browse workers with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkerCard"

  /workers/categories:
    get:
      operationId: listCategories
      tags: [workers]
      summary: Get all available skill categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /workers/{workerId}:
    get:
      operationId: getWorkerProfile
      tags: [workers]
      summary: Get full worker profile with reviews
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for the current user (customer sees their requests, worker sees incoming requests)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request (customer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{bookingId}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Accept, reject, or complete a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking (customer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getPlatformStats
      tags: [stats]
      summary: Get platform-wide stats (total workers, bookings, top categories)
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    SignupBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        city:
          type: string
        userType:
          type: string
          enum: [customer, worker]
        skill:
          type: string
          description: Required for workers
        hourlyRate:
          type: number
          description: Required for workers (PKR)
        bio:
          type: string
      required: [name, phone, password, city, userType]

    LoginBody:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
        userType:
          type: string
          enum: [customer, worker]
      required: [phone, password, userType]

    UserInfo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        userType:
          type: string
          enum: [customer, worker]
        workerId:
          type: integer
          description: Present for workers
      required: [id, name, phone, city, userType]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
      required: [user]

    WorkerCard:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        city:
          type: string
        skill:
          type: string
        hourlyRate:
          type: number
        avgRating:
          type: number
        totalReviews:
          type: integer
        bio:
          type: string
      required: [id, userId, name, city, skill, hourlyRate, avgRating, totalReviews]

    Review:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        workerId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required: [id, bookingId, customerId, workerId, rating, createdAt]

    WorkerProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        city:
          type: string
        skill:
          type: string
        hourlyRate:
          type: number
        avgRating:
          type: number
        totalReviews:
          type: integer
        bio:
          type: string
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
      required: [id, userId, name, city, skill, hourlyRate, avgRating, totalReviews, reviews]

    CreateBookingBody:
      type: object
      properties:
        workerId:
          type: integer
        date:
          type: string
          description: ISO date string
        time:
          type: string
          description: e.g. 09:00
        address:
          type: string
        notes:
          type: string
      required: [workerId, date, time, address]

    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected, completed]
      required: [status]

    Booking:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        workerId:
          type: integer
        workerName:
          type: string
        workerSkill:
          type: string
        workerHourlyRate:
          type: number
        date:
          type: string
        time:
          type: string
        address:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, completed]
        hasReview:
          type: boolean
        createdAt:
          type: string
      required: [id, customerId, workerId, date, time, address, status, hasReview, createdAt]

    CreateReviewBody:
      type: object
      properties:
        bookingId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [bookingId, rating]

    PlatformStats:
      type: object
      properties:
        totalWorkers:
          type: integer
        totalBookings:
          type: integer
        totalCategories:
          type: integer
        topCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
        cities:
          type: array
          items:
            type: string
      required: [totalWorkers, totalBookings, totalCategories, topCategories, cities]